feat: security/correctness audit fixes, X-style comments, WebSocket push, admin UI, cookie auth, PG adaptation
- Search fix: LIKE ESCAPE '!' (dialect-neutral) replaces the '\' form that
  broke under the MySQL-dialect lexer (H2 MODE=MySQL)
- Security: PBKDF2 hashing with legacy auto-upgrade, constant-time HMAC
  compare, JWT placeholder fail-fast, OAuth state cookie + origin-pinned
  redirect, image-only uploads with sanitized names + nosniff, profile
  sanitization, rate limiting, typed-body validation, security headers
  filter (CSP/XFO/nosniff/Referrer-Policy), HttpOnly SameSite=Lax session
  cookie replacing localStorage JWT
- Correctness: last_comment_time maintenance, counter consistency
  (comment/topic delete & create), deleted topics 404, partial article
  updates, transactional like/follow/favorite, unique composite indexes,
  pagination clamps and long offsets, 400s instead of 500s
- Features: X-style comment replies (quoteId + quote preview + auto-@),
  WebSocket notification push (/ws/notify), admin panel UI with
  requiresAdmin guard, config-driven admin bootstrap, audit logs
  (operate/download), JSON message extraData
- Data layer: 13 query indexes, dead tables pruned from ALL_ENTITIES,
  LONGTEXT->TEXT, H2 switched to MODE=PostgreSQL, production target
  switched from MySQL to real PostgreSQL (postgresql JDBC driver added,
  verified against PG 18.4: 21 tables + 20 indexes + full smoke)
- Tests: AuthService/UserService/CommentService suites (19 total) +
  tools/smoke-api.sh (16 checks); CI workflow; ESLint/Prettier config

+++ point-service/src/main/java/com/jujin/point/service/MessageService.java
@@ -34,7 +34,7 @@ public class MessageService {
     }
 
     public List<Message> getUserMessages(long userId, int page, int pageSize) {
-        int offset = (page - 1) * pageSize;
+        long offset = (long) (page - 1) * pageSize;
         return db.query(
             "SELECT * FROM bbs_message WHERE user_id = ? ORDER BY create_time DESC LIMIT ? OFFSET ?",
             userId, pageSize, offset).list(Message.class);
